MYR-176 Split dispatcher.go under the 300-line rule (retry/reconcile/helpers files)
