+++ internal/dispatch/dispatcher_helpers.go
@@ -1,0 +1,14 @@
+package dispatch
+
+// redactVIN shows only the last 4 chars of a VIN for logs (empty stays empty).
+func redactVIN(vin string) string {
+	if len(vin) <= 4 {
+		return vin
+	}
+	return "***" + vin[len(vin)-4:]
+}
+
+// discardWriter drops log output for the nil-logger default.
+type discardWriter struct{}
+
+func (discardWriter) Write(p []byte) (int, error) { return len(p), nil }
